chore: Move ConversationMessageTimerCellDescription into its own file
